Reduced COMMAND.COM to 14340 bytes.  Still linking C malloc and free, which may be taking large number of bytes.  However, my code is not using C malloc, and cannot find which C-runtime function is referencing malloc and free at this time.

+++ FMT_DOS/COMMAND/UTIL.C
@@ -496,7 +496,7 @@ char *FindRedirection(char *foundChar,char arg[])
 
 	return NULL;
 }
-void Itoa(char *a,unsigned long int i)
+void Itoa(char *a,unsigned int i)
 {
 	char rev[16];
 	if(0==i)
